requirements.txt
alias_free_torch==0.0.6
deepspeed==0.16.5
descript_audio_codec==1.0.0
einops==0.8.1
einops_exts==0.0.4
encodec==0.1.1
g2p_en==2.1.0
huggingface_hub==0.30.2
k_diffusion==0.1.1.post1
local_attention==1.11.1
matplotlib==3.10.1
numpy==2.2.4
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
regex==2024.11.6
Requests==2.32.3
safetensors==0.5.3
scikit_learn==1.6.1
scipy==1.15.2
timm==1.0.15
tqdm==4.67.1
transformers==4.51.3
vector_quantize_pytorch==1.22.12
wandb==0.19.9
webdataset==0.2.111
x_transformers==2.2.11
